Bug 491301 - Java/Build Path/Classpath Variable pref page has wrong colors
Non editable entries are white text on black background on Linux with
Gtk default theme Adwaita (due to usage of COLOR_INFO_* constants, which
are ment for popups) which is way too intrusive in the UI.
Additionally when using Eclipse dark theme having same colors makes it
undistinguishable from the rest of the UI using dark pallette.
Make the class not implement IColorProvider as this should be handled by
the theme.
Remove useless non-javadoc comments while at it.

Change-Id: If84b6119c69250fc98cef7ef31e5e3240d3a3edb
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>