Bug 519950: Improve Dark Theme so it works well with Menlo Font
The dark theme should not use bold font face massively. This is
not mac specific but surfaced when we switched the default
text font on mac from Monaco to Menlo. Monaco does only contain
a "regular" font face so all "bold" word were "regular". So this
issue was not visible on macOS. On other platforms this issue
should have been visible.

Also change some colors as some of them were too close the 
keyword-color.

Change-Id: Icc3f73a9daf03c9544cc839ef50ed5f4b13910f7
Also-by: Till Brychcy <register.eclipse@brychcy.de>
Signed-off-by: Matthias Becker <ma.becker@sap.com>