Bug 526010 - Very long check when deleting thousands of sub-packages
When deleting a package with a few thousand sub-packages, a check prior
to removal takes up to several minutes. This check cannot be cancelled
by canceling the delete operation. 

This change introduces various improvements to the running time of the
check, mostly without extra memory cost. The check for a few thousand
sub-packages then takes a few seconds instead of minutes. It also add a
progress monitor to the operation.

Change-Id: I978b928b73953dd0c42447a1e13eb33f13479e84
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>