Bug 530125 - Solve Performance - Method invokes inefficient Number constructor; use static valueOf instead in JDT UI
Use search and replace to replace
* new Integer(
* new Float(
* new Double(
* new Long(

with the corresponding valueOf( calls.

The following source files are NOT changed to avoid test failures:

    org.eclipse.jdt.ui.tests.quickfix.AssistQuickFixTest
    org.eclipse.jdt.ui.tests.quickfix.CleanUpStressTest
    org.eclipse.jdt.junit.tests.LegacyTestRunListenerTest
    org.eclipse.jdt.junit.tests.TestRunListenerTest3

Also the test data for new Integer was untouched in case that cased a
test failure.
This way we still perform the relevant tests for the usage of these
deprecated (as of Java 9) constructors.

Change-Id: Idf7293a06635d530704f5b59226e594f44fafe6d
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>