Bug 38201: [quick fix] Allow creating abstract method
This change add another quick fix proposal which allows the creation of
an abstract method if the target type is a class and is abstract.

An existing unit test was amended to include this new functionality.
A new test was added to test abstract parent classes in a nested case.

Change-Id: I0a4baee73ff79889f4c8e2e754ca9a94555f804b
Signed-off-by: Jens Reimann <jreimann@redhat.com>