Bug 539470 - Eclipse freezes on opening packages in Project Explorer
The NonEssentialElementsFilter applies the complete filter logic to all
children, recursively. But it only needs to know whether there is at
least one child selected by the existing filters. Therefore rework the
implementation to use an early exit after the first hit.

Runtime of the filter for the example attached to the bug drops from 24
to 2.6 seconds and the main thread is blocked much less when expanding
the huge package hierarchy in the package explorer.

Change-Id: I0076429db5505cc84dda17e49afc0e29799322aa
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>