Bug 539289 - "Surround With try/catch" should revert "var" initializer.
- "Surround With try/catch" moves the variable initializer into the "try"
while creating a null initializer declaration outside. However, it is a
compile-time error for the local variable type to be "var", with a null
initializer. The "var" type must be changed to the inferred type.
- When "var" must be replaced with the inferred type, the quick fix
label should describe the additional steps taken
- Create helper ASTNodes.isVarType(..) to determine if a given
ASTNode is either a SimplNameVariable or DeclarationStatement whose type
is 'var'
- Use ASTNodes.isVarType(..) in
QuickAssistProcessor.getConvertVarTypeToResolvedTypeProposal(..)

Change-Id: I63d315afdf5f86ddfb5c85dacebafcc956a8c6c7
Signed-off-by: Roland Grunberg <rgrunber@redhat.com>