Bug 526339 - [quick fix] Add Service Provider Method.
Offer quick fixes that create a "provider" method inside the service
provider with a return type that is either the same, or a subtype of the
service. Subtypes of the service provider are offered in linked mode so
as to defer subtype hierarchy computation.

Take this action under the following circumstances :

    - When a java module source file provides an interface with an
      invalid service provider
    - When a java module source file provides an interface whose service
      provider lacks a public no-arg constructor

- Add testcase for checking the proposal, and resolution
- Added testability for linked mode proposals

Change-Id: Ib80f4f541f36de5d561361238556c32456b64b6f
Signed-off-by: Roland Grunberg <rgrunber@redhat.com>
Also-by: Stephan Herrmann <stephan.herrmann@berlin.de>