Bug 451777 - Defer JREBlock initialization to background jobs
Initializing the JRE information can take longer as expected and leads
to UI freezes. To avoid the freeze the JRE information is fetched in a
background thread and once finished, the information is filled into the
fields of JREBlock. The JRE fields are kept disabled until JREBlock is
initialized.

The Validator disallows continuation of the wizard when one of the
combos is selected, but has no entries yet as they are fetched ATM.

Added some disposal checks that are required to avoid access of widgets
when the dialog is closed while the JVM installations are fetched.

Change-Id: I5c1b51f18d6e4d3b407ec6ccdc75f2f67df89896
Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>