Bug 541170 - Incorrect background color for module source code
When needing to render module content, we create an entirely new
SimpleJavaSourceViewerConfiguration with special property simply to
activate code in JavaCodeScanner (create9ModuleInfoRules) which
understands modules.

configure() seems to cause problems rendering the background, but the
only reason it is being called is to re-initialize the code scanner
mentioned above so it understands module rules. Therefore, we can simply
call :

JavaSourceViewerConfiguration.getPresentationReconciler(ISourceViewer)

and effectively update the code scanner without calling configure.

Change-Id: I9cbf8c3774e0d8fe3107f19e6c20eb62776615ac
Signed-off-by: Roland Grunberg <rgrunber@redhat.com>