Bug 526339 - Reverse compilation unit comparison to avoid NPE.
When constructing the valid linked-mode proposals for the provider
method's return type, we compare the compilation unit where the method
will reside against the compilation unit of the candidate return type to
evaluate visibility. If a candidate return type is a binary type, it
will have no compilation unit (null) so we should reverse the
comparison.

- Modify testBasicNewServiceProvider to test the case of a service being
an existing type

Change-Id: I2cf06a5579075f3b706cde2e1dd8f728fac33dba