Bug 545261 - [12][quick assist] Split case expressions/statements.
Locate the statement index of the selected node's SwitchCase and copy
all its statements (they'll be the subsequent statements up to and
excluding the next SwitchCase), inserting them before the original
SwitchCase along with the newly created SwitchCase representing a single
label. This is done for every label of the original SwitchCase.
Afterwards the original SwitchCase and its statements are removed.

- This applies to JLS12 or higher, where the selected ASTNode is the
'case' keyword (eg. SwitchCase) or a label of the 'case' expression (eg.
element whose parent is a SwitchCase)
- The ordering of the SwitchCase labels is preserved
- Add AssistQuickFixTest12 for testing this quick assist

Change-Id: I3fd37046b63622174a46ff20ebff43205b5d18ea