Bug 475944 - Removes usage of IPluginDescriptor in JDT
IPluginDescriptor will be deleted in 4.12

Activate the tests, to avoid flakiness the same list from
PluginsNotLoadedTest is specified as plug-ins which should not be
activated and the plug-ins which are activated by opening an Java editor
have been removed from the list. This is more stable compared to the
definition of a list of plug-ins which should be activated, as this
depends on the runtime which is used.

Change-Id: I70ed32ad6a9c4cf59f45ede231ab9ad11a6784d2
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>
Signed-off-by: Roland Grunberg <rgrunber@redhat.com>