Fix bug 488942: [extract method] Extract method fails to handle side effect in loops
- Add re-entry analysis for nested loops in InputFlowAnalyzer.
  This works for all loop types (for, enhanced_for, do and while loops.)
- Add a test "A_test735" for the bug that is fixed.

Before only iterating the outermost loop was analyzed. Therefore it was
not discovered that variables changed in the extracted code inside a
nested loop could be read in the next iteration of the inner loop.

For variables declared outside the outermost loop this was saved by the
re-entry analysis of the outermost loop, as the changed variable is also
read in the re-entry of the outermost loop (in the initial iteration of
the nested loop).
Therefore the end result was OK, even though it did not pick up the read
of the variable on re-entry of the nested loop.

However for variables declared inside the outermost loop the variable
goes out of scope and a new instance is used in the re-entry of the
outermost loop. Therefore only a re-entry analysis of the nested loop
itself can discover that the changed variable is read.

Change-Id: I1152af0c52a6ac01e71a7d5e5a0aa079b4a777a6
Signed-off-by: Dan N. Christensen <dan@celticdanes.net>