Bug 546793 - [code mining] parameter names shifting offset
The code minings can be redrawn with data based on a state that doesn't
match the actual viewer content because of a change since code mining
update was triggered.
This cancels the code mining update in such case, to avoid erroneous
redraw, and only rely on usual and generic annotation shifting in text
viewer (that works well).

Change-Id: I4f370f837ccb9df2749f4fcac46fa17ce12673d3
Signed-off-by: Mickael Istria <mistria@redhat.com>