Bug 549457 - Autobuild does nothing on org.eclipse.jdt.core.prefs change
If the file .settings/org.eclipse.jdt.core.prefs changes, JDT does not
automatically rebuild the project. E.g. pulling different settings from
git will not result in a re-build; only changing the settings from the
Java compiler preference page will ask the user if they wish to rebuild
the workspace.

After a change in JDT core, a full build is triggered when such changes
are done. The preference page dialog option to not build on applying
Java compiler preference becomes invalid; this change removes this
dialog option.

In case this new behavior causes problems, e.g. in tooling which
re-generates project files without actually changing them, it is
possible to return to the previous behavior with the VM property:

-Dorg.eclipse.jdt.core.disableAutoBuildOnSettingsChange=true

Change-Id: Ief66752f9319051698209fbd28a5e9ca6bfee400
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>