Bug 549473 - [13] Text Block support in Java Editor for JEP 355 Text Block
This patch contains changes as below:
1) IJavaPartitions introduces a new partition pertaining to Text Block.
2) FastJavaPartitionScanner to handle Text Blocks when required.
3) FastJavaPartitioner is a new class extended from FastPartitioner so
that the partitions can be recalculated when the preference value of
preview enabled feature changes.
4) PreviewFeaturesSubProcessor provides 2 new utility functions
pertaining to preview features and project specific settings.
5) JavaMultiLineStringAutoIndentStrategy to handle Text Block auto
indentation.
6) IJavaPartitionerManager handles the notifications to the
IDocumentPartitioners upon the change in preferences.
7) JavaPartitionerManager implements IJavaPartitionerManager and also
maintains the Editor info that is required by the FastJavaPartitioner to
find the Project information required. It also maintains an
Editor-Partitioner map to cleanup the editor info once the editor is
discarded.

TODO: Add test cases for the Text Blocks. This will be done as part of
another patch for this bug.

Change-Id: I7144e0462d52934c73f1644d0c2bab18fad880d5
Signed-off-by: Kalyan Prasad Tatavarthi <kalyan_prasad@in.ibm.com>