Bug 550451 - [quick assist][rulers] Quick Assist light bulb no longer works
This was caused because the AssistAnnotation created here is of type
"org.eclipse.jdt.ui.warning" but has no preference.But there is a check
added to get the preference. Added code to not check for preference for
AssistAnnotation. Also modified few Quick Assist proposals so also to 
return true only if a proposal exists.

Change-Id: I08b71455a2bddd4a749937f9dbcbe8ba544bb9f7
Signed-off-by: Kalyan Prasad Tatavarthi <kalyan_prasad@in.ibm.com>