Bug#553608 - Only consider direct parent type for some rules
Some rules searched recursively for a TypeDeclaration in the parent
hierarchy. This lead to the case that some modifiers were removed from
nested fields / methods that should only be removed when they are owned
by an interface. When the owner itself is an enum in a nested interface
this retrieved the interface declaration when searching for the type,
and did not consider that the owner is an enum.

Changed to consider only the direct parent of a field/method/enum
declaration for rules.

Added test case to cover the regression.

Change-Id: Ia6cc6de4dc21bf2b9f1b8ea4f7db5229ae4c6099
Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>