Bug 558313 - get rid of static initializations
Remove ClassRules: Though they sometimes can be faster they can be kind
of evil as they might be executed not at the right point in time or even
more than once depending on the test environment.
Try out if this is running through.
Similar problems can be caused by "static { }" parts or class variables
affecting the test. Unfortunately we have some of these.

Change-Id: Ief9b67df13c0942759f3796f57d18b1da919c47f
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>