Bug 561986 - [chain] Fix proposal inserted with no explicit invocation
Currently, a range of (invocationOffset, invocationOffset + tokenLength)
is used for defining the tokens to be replaced when a chain completion
proposal is applied. This range fails for completions that don't use an
explicit invocation trigger character. Instead, use a range of
(tokenStart, tokenStart + tokenLength) to correctly define the tokens to
be replaced.

- Add testcase

Change-Id: Ic403df42f0750de05b39477acf055b6afb42fb1b