Use switch rather than if/else if/else where possible
Change cascades of ifs which can be converted to switch. A
switch statement might be faster than an if-then-else chain. And it
improves clarity. The problem with if..else chain is that I have to look
into all the if conditions to understand what the program is doing. And
the variable might change in the chain processing.

Change-Id: Idc1d417a4ee4080c23b589fe98cb1859c6887f14
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>