Bug 548430 - [Cocoa] NullPointerException in Image.internal_new_GC (4.12.0RC2)
- There is an underlying assumption while creating breadcrumb arrows.
The image creation logic assumes the new image created is always created
with transparent color.

Currently only the opaque areas are set with alpha 255, which is opaque.
But the areas where transparency should be set is not handled at all.

This patch sets transparent area with alpha 0, which is transparent.

Change-Id: I95c7b0baf3870d1890a4b9ff3fd704837aea4b86
Signed-off-by: Sravan Kumar Lakkimsetti <sravankumarl@in.ibm.com>