Bug 562289 - add junit sample project sources to git
While the sources are now a real project within the jdt.ui git
repository it is possible to provide a maven-assembly-plugin
configuration to assemble the distribution package. This way it can be
consumed within

org.eclipse.jdt.ui.examples.projects

from this on-the-fly created jar instead or additionally to the static
jar file we have already as

junit381src.jar

From now on all updates in this project automatically are part of the
junit sample project and it is much easier to maintain it.

Change-Id: Icf240050d4048d4b161037597ffec339b0fe03df
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>