Bug 567150 - no hidden rules any longer
By nesting construction make sure that only one rule per test class is
instantiated. Unfortunately the code this way is not really nice.
However as long as there is no better solution this way we get rid of
some warnings and a potential source of problems regarding test
execution cleanup.

MoveInnerToTopLevelTests.java contains now a sleep to make the test run.
The reason for the need of this sleep is not known yet. It is needed
since I changed its @ClassRule to @Rule.

Change-Id: Iae3f70d853cf6703b564c218c69fccef38222d5b
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>