Bug 567084 - use null check instead instanceof where possible
This instanceof test will always return true (unless the value being
tested is null). Although this is safe, make sure it isn't an indication
of some misunderstanding or some other logic error. If you really want
to test the value for being null, perhaps it would be clearer to do
better to do a null test rather than an instanceof test.

Change-Id: I9ea2f13cdab94e1ba5f36c54351e94fe0623f6d6
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>