Bug 567483 - explicitly call doOpenExternalJarFileDialog() with null
This makes spotbugs stop to complain. The null case is used in fact.

Load of known null value in
org.eclipse.jdt.internal.ui.preferences.UserLibraryPreferencePage.openJarFileDialog(CPListElement,
Object)

The variable referenced at this point is known to be null due to an
earliercheck against null. Although this is valid, it might be a mistake
(perhaps you intended to refer to a different variable, or perhaps the
earlier check to see if the variable is null should have been a check to
see if it was non-null).

Rank: Of Concern (16), confidence: Normal
Pattern: NP_LOAD_OF_KNOWN_NULL_VALUE
Type: NP, Category: STYLE (Dodgy code)


Change-Id: Id362c0862051865d598ab588fc3fbb6bc8a12bc2
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>