Bug 567483 - Don't check again for null if it is clear
Load of known null value in
org.eclipse.jdt.internal.ui.text.AbstractInformationControl$3.mouseMove(MouseEvent)

The variable referenced at this point is known to be null due to an
earliercheck against null. Although this is valid, it might be a mistake
(perhaps you intended to refer to a different variable, or perhaps the
earlier check to see if the variable is null should have been a check to
see if it was non-null).

Rank: Of Concern (16), confidence: Normal
Pattern: NP_LOAD_OF_KNOWN_NULL_VALUE
Type: NP, Category: STYLE (Dodgy code)


Change-Id: If1f4208d115de066800fe7fe26a4d9d3457e34dc
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>