Bug 567483 - Don't use variable known to be null
Spotbugs logs it as error

Load of known null value in
org.eclipse.jdt.internal.ui.wizards.buildpaths.LibrariesWorkbookPage.openJarFileDialog(CPListElement)

The variable referenced at this point is known to be null due to an
earliercheck against null. Although this is valid, it might be a mistake
(perhaps you intended to refer to a different variable, or perhaps the
earlier check to see if the variable is null should have been a check to
see if it was non-null).

Rank: Of Concern (16), confidence: Normal
Pattern: NP_LOAD_OF_KNOWN_NULL_VALUE
Type: NP, Category: STYLE (Dodgy code)


Change-Id: I5a9892425331a8f4f603ad82336b7ae011023338
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>