Bug 567483 - Load of known null value
org.eclipse.jdt.internal.core.manipulation.dom.ASTResolving.findParentBodyDeclaration(ASTNode,
boolean)

The variable referenced at this point is known to be null due to an
earliercheck against null. Although this is valid, it might be a mistake
(perhaps you intended to refer to a different variable, or perhaps the
earlier check to see if the variable is null should have been a check to
see if it was non-null).

Rank: Of Concern (16), confidence: Normal
Pattern: NP_LOAD_OF_KNOWN_NULL_VALUE
Type: NP, Category: STYLE (Dodgy code)


Change-Id: I98d5eb443b29ccd638c1c65e991966eebca86581
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>