Bug 566507 - [15] Problems in renaming canonical constructor parameter
This code change contains support for renaming of canonical constructor parameters of a record. Also the support is used to provide to check for a duplicate local variable. Tests have been added and modified for the same.

Point to note is that now, selecting a Canonical Constructor parameter, or accessor method in a record calls rename of the record component.

Change-Id: I8409238698e24314744544d75fdabdf8bc3a385c