Bug 568822 - Expressions with two repeated operands should not be refactored
Let be the following case:
(X && Y) || (!X && !Y)

If you run both "Ternary operator" and "Use '^' or '==' on booleans"
rules, you don't know if you will get this:
(X ? Y : !Y)

...or this:
X == Y

...because both rules match the same code. They are in conflict and the
result is a random result.

So we should disable the matching for "Ternary operator" rule because
the result is worse.

Change-Id: I84a2e436060942998a9d1d1677129acc28a16a9d
Signed-off-by: Fabrice Tiercelin <fabrice.tiercelin@yahoo.fr>