Bug 566379 - Flush testEnded messages to client
The testEnded and testFailed messages weren't being flushed back to the
client, so it was up to a subsequent message send to flush the queue.
This resulted in incorrect numbers being reported for test cases, since
the time displayed in the UI is measured from when the testEnded message
was received from the test runner.

By flushing the messages as they are created, we ensure that the client
can calculate reliable times for test runs, such as when a JUnit 4
BeforeClass or a JUnit 5 BeforeAll annotation is used.

Change-Id: Ife4ee1671f3c28df45f6d3f1f82d036951a57dd4
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>