Bug 570773 - [cleanup] "Source fixing" new tab page
The tab content (at this time):

Code standardization
 - StandardComparisonCleanUp
 - BitwiseConditionalExpressionCleanup

It is planned to add cleanups that avoid null pointers. Those cleanups
may trigger zombie code. A zombie code is a dead code that is dead
because an error occurs before. The day someone fixes the error, the
zombie code comes back to life and alters the behavior.

Change-Id: Ife9d745ee669ad4c330614412a8a6c0b2abd4432
Signed-off-by: Fabrice Tiercelin <fabrice.tiercelin@yahoo.fr>