Bug 572323 - [build] Make "Autoboxing" and "Unboxing" cleanup tests run faster
We are experimenting slowness on the Jenkins builds. This ticket
rewrites the tests for "Autoboxing" and "Unboxing" cleanups. It merges a
dozen of tests into two ones. They still cover the whole features. Now
they are written the same way as the tests of the other cleanups.

Also remove warnings.

Change-Id: If1115241dd41432f973833e6c171e7121811f48c
Signed-off-by: Fabrice Tiercelin <fabrice.tiercelin@yahoo.fr>