Bug 572352 - [save action] Reduce SaveParticipantTest failures
We often face random SaveParticipantTest failures. This ticket does not
fix the cause of the problem but reduce the probability of occurrence of
such a crash.

It seems that the save action options are changed by a concurrent thread
during the tests so this ticket brings the writing and the reading of
this options closer. Other computations are done before.

To reduce the timing, the StringBuffer usage have been replaced by a
unique String. Indeed, a concatenation of string literals is compiled as
one single big string literal.

Change-Id: Ife33362a296984be7c8b6c49d2f8b113d11c26be
Signed-off-by: Fabrice Tiercelin <fabrice.tiercelin@yahoo.fr>