Bug 572495 - [cleanup] "Simplify lambda expression and method reference syntax" should handle type ref for self class
For an unknown reason, MethodInvocation.resolveMethodBinding() seems to
fail when the method is defined in the class we are. JDT Core should
probably be fixed for that. Here is a workaround for the cleanup.

Given:
Function<String, Integer> function= numberInText ->
E.dummy(numberInText);

When:
Applying "Simplify lambda expression and method reference syntax"
cleanup...

Then:
Function<String, Integer> function= E::dummy;

Change-Id: I90f678686207f52bf81b7096310e1bd83fbb0541
Signed-off-by: Fabrice Tiercelin <fabrice.tiercelin@yahoo.fr>