Bug 574376 - [dogfooding][cleanup] Use new StringBuffer -> StringBuilder cleanup for test code
Runs the new StringBuffer-> StringBuilder

This help to determine that it works correctly for clients we should run
it also on our own code base. This help to ensure that the cleanup works
fine and helps us to cleanup our own code base.

From the commit message of the cleanup:

- create a cleanup to replace usage of StringBuffer with
  StringBuilder.  The two are equivalent, but StringBuffer is
  thread-safe and synchronized which makes it slower than
  StringBuilder which was introduced in Java 1.5.


This uses the local variable change only.

Change-Id: I6e4de0112ab8183243d80c4a34c5281529d9f170
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.ui/+/182328
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>