Bug 574640 - New Java Project dialog: show scrollbars when required
In the "New Java Project" dialog/wizard, on the first page, show
scrollbars when required instead of skipping the content at the bottom.
Otherwise, when the dialog height is too small, the "Create
module-info.java file" checkbox and any existing infos and warnings
displayed below the checkbox may be hidden and thus become inaccessible.

Change-Id: I1423b621f339f605ac0c56a655bbda09da9230a3
Signed-off-by: Holger Voormann <eclipse@voormann.de>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.ui/+/182740
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Kalyan Prasad Tatavarthi <kalyan_prasad@in.ibm.com>