Bug 578491 - Switch JRE to another JRE results in re-ordering
This change preserves classpath order on JRE version changes during
BuildPathBasePage.moveCPElementAcrossNode(). In addition, a small
optimization is added to the first loop of moveCPElementAcrossNode() -
the loop is exited when the root element of the selected element is
found.

As is, for a move from modulepath to classpath (e.g. change JRE 11 to
JRE 8), the method moves children of the root container for modulepath
to classpath. Additionally, it removes the current JRE container, and
adds the new JRE container. This breaks classpath order in two ways.

First, the order of children in the root container is changed, due to
the removal and addition of the JRE container (if it was at first
position, now its at last position).

And second, LibrariesWorkbookPage.updateClasspathList() is unable to
preserve order of the classpath due to first being invoked with no
classpath containers in any root container. To then be invoked with the
correct elements in the root containers. The method loses track of the
correct positioning of the classpath containers.

The above is true also when moving from classpath to modulepath (e.g.
change JRE 8 to JRE 11).

Change-Id: I2ac18860deb420da611bf3525eead6b3ae683e24
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.ui/+/190227
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Kalyan Prasad Tatavarthi <kalyan_prasad@in.ibm.com>