Refactor rename wrongly reports shadowings#201
During a refactor rename, its possible that new occurrences of the
renamed methods are not correctly detected by
RenameMethodProcessor.batchFindNewOccurrences().

This is due to batchFindNewOccurrences() creating a scope based on the
set of new (renamed) methods, using
RefactoringScopeFactory.create(IMember[]). This create() method
essentially picks the project of the first of the specified IMember
objects, at least for trivial refactor renames. However that project
might not be sufficient to search for all new renamed method
occurrences, it might instead be a project that is a leaf dependency and
not a root dependency.

This change adjusts the used scope to be that of the project scopes of
all new methods provided to batchFindNewOccurrences(), instead of just
using the first method.

Fixes: #201