Show option to create a new java project in the project explorer (#362)
Currently if one opens the "Package Explorer" and right-click an item the
user is offered to create a new java project. On the other hand, in the
"Project Explorer" this option is not offered what is confusing.

This adds a commonWizard with a filter for the active perspective that
shows this option in the "Project Explorer" as well.