remove author from default code templates (#374)
This was useful at times where many people did not have version control
or a version control system that was not easy to use and query. Today
that leads to lots of problems:

* The same information is better available in the version control
history, with attribution by line instead of assuming this one author
for every line currently in the file
* People copy/paste classes to create "something similar", without
fixing this comment.
* It may leak information of the private user account to the public (at
least on Windows, this defaults to your local user account id, and my
company admin would surely not like me posting that account id
everywhere).
* Even developers without an underlying version control system should
not rely on this information, but rather be nudged to set up a real
version control system.