Always use brackets in blocks of code by default #667
Change the default settings of the built-in clean-up profile and in the
"Additional actions" of the "Save Actions"; set them to always use
brackets for blocks of code (even for one-liners).

These settings are just the defaults, they can be overwritten.

Fixes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/667