Replace o.e.o.s.resolver.VersionRange by org.osgi.framework.VersionRange
The class org.eclipse.osgi.service.resolver.VersionRange belongs to the
outdated org.eclipse.osgi.service.resolver package and already inherits
org.osgi.framework.VersionRange.

Additionally in P2Utils.findBundle() the check if the bundle file exists
is deferred to the latest possibility in order to speed-up the search if
multiple versions are present and avoids a NPE if the bundle-file cannot
be determined.