Dont show children for JUnit 5 @TestTemplate tests that ran only once
When running JUnit 5 @TestTemplate tests, extra test suites are shown in
the JUnit view to represent the tests in the template. Those suites
introduce unnecessary noise when the template contains only one test -
e.g. when repeating tests that failed.

This change adjusts TestSuiteElement.getChildren(), pruning each
TestSuiteElement that has a single TestCaseElement child that is also a
dynamic test.

Fixes: #945
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>