Quickfix Converting to Enhanced For Loop Creates Invalid Code #1008
The change fixes the conversion of an iterator based for loop to an enhanced for loop. When there is an iterator.next() in the body of the for loop and is a direct only child of the whole statement, it becomes a standalone identifier (var;) on conversion which is an errorful statement in Java.

Fixes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/1008