Ensure "Make Static" refactoring only adds parameter if necessary #1044
The "Make Static" refactoring currently adds a "this" parameter to the
method in cases in which it is actually not necessary. This is
particularly the case the body of the method to be refactored contains
invocations of methods on a newly instantiated object. The according
logic erroneously considers such a method invocation as a reason to add
a parameter.

With this fix, the identification of whether a parameter needs to be
added to the method depends on whether the logic calculating an actual
change to the method body is actually applied. An according
regression test for the behavior is added.

Fixes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/1044